+++ madsim-etcd-client/src/bytes.rs
@@ -18,7 +18,7 @@ impl Bytes {
 
 impl Debug for Bytes {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "\"{}\"", self)
+        write!(f, "\"{self}\"")
     }
 }