+++ madsim-rdkafka/src/sim/producer.rs
@@ -121,10 +121,11 @@ where
             .map_err(|e| KafkaError::ClientCreation(e.to_string()))?;
         let config: ProducerConfig = serde_json::from_str(&config_json)
             .map_err(|e| KafkaError::ClientCreation(e.to_string()))?;
-        let addr = config
-            .bootstrap_servers
-            .parse::<SocketAddr>()
-            .map_err(|e| KafkaError::ClientCreation(e.to_string()))?;
+        let addr: SocketAddr = madsim::net::lookup_host(&config.bootstrap_servers)
+            .await
+            .map_err(|e| KafkaError::ClientCreation(e.to_string()))?
+            .next()
+            .ok_or_else(|| KafkaError::ClientCreation("invalid host or ip".into()))?;
         let p = BaseProducer {
             _context,
             config,