+++ madsim-rdkafka/tests/test.rs
@@ -1,6 +1,7 @@
 #![cfg(madsim)]
 
 use futures_util::StreamExt;
+use madsim::net::NetSim;
 use madsim::runtime::Handle;
 use madsim_rdkafka::{
     admin::*,
@@ -21,6 +22,8 @@ use std::{
 async fn test() {
     let handle = Handle::current();
     let broker_addr = "10.0.0.1:50051".parse::<SocketAddr>().unwrap();
+    NetSim::current().add_dns_record("broker", broker_addr.ip());
+
     handle
         .create_node()
         .name("broker")
@@ -38,7 +41,7 @@ async fn test() {
         .build()
         .spawn(async move {
             let admin = ClientConfig::new()
-                .set("bootstrap.servers", broker_addr.to_string())
+                .set("bootstrap.servers", "broker:50051")
                 .create::<AdminClient<_>>()
                 .await
                 .expect("failed to create admin client");
@@ -60,14 +63,14 @@ async fn test() {
         .build()
         .spawn(async move {
             let producer = ClientConfig::new()
-                .set("bootstrap.servers", broker_addr.to_string())
+                .set("bootstrap.servers", "broker:50051")
                 .create::<BaseProducer>()
                 .await
                 .expect("failed to create producer");
 
             // generate an element every 0.1s
             for i in 1..=30 {
-                let key = format!("1.{}", i);
+                let key = format!("1.{i}");
                 let payload = [i as u8];
                 let record = BaseRecord::to("topic").key(&key).payload(&payload);
                 producer.send(record).expect("failed to send message");
@@ -85,14 +88,14 @@ async fn test() {
         .build()
         .spawn(async move {
             let producer = ClientConfig::new()
-                .set("bootstrap.servers", broker_addr.to_string())
+                .set("bootstrap.servers", "broker:50051")
                 .create::<BaseProducer>()
                 .await
                 .expect("failed to create producer");
 
             // generate an element every 0.2s
             for i in 1..=30 {
-                let key = format!("2.{}", i);
+                let key = format!("2.{i}");
                 let payload = [i as u8];
                 let record = BaseRecord::to("topic").key(&key).payload(&payload);
                 producer.send(record).expect("failed to send message");
@@ -114,7 +117,7 @@ async fn test() {
         .build()
         .spawn(async move {
             let consumer = ClientConfig::new()
-                .set("bootstrap.servers", broker_addr.to_string())
+                .set("bootstrap.servers", "broker:50051")
                 .set("enable.auto.commit", "false")
                 .set("auto.offset.reset", "earliest")
                 .create::<BaseConsumer>()
@@ -146,7 +149,7 @@ async fn test() {
         .build()
         .spawn(async move {
             let consumer = ClientConfig::new()
-                .set("bootstrap.servers", broker_addr.to_string())
+                .set("bootstrap.servers", "broker:50051")
                 .set("enable.auto.commit", "false")
                 .set("auto.offset.reset", "earliest")
                 .create::<StreamConsumer>()